Moved function to common
