+++ pkg/network/protocols/common.go
@@ -1,0 +1,20 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+package protocols
+
+// below is copied from pkg/trace/stats/statsraw.go
+// 10 bits precision (any value will be +/- 1/1024)
+const roundMask uint64 = 1 << 10
+
+// NSTimestampToFloat converts a nanosec timestamp into a float nanosecond timestamp truncated to a fixed precision
+func NSTimestampToFloat(ns uint64) float64 {
+	var shift uint
+	for ns > roundMask {
+		ns = ns >> 1
+		shift++
+	}
+	return float64(ns << shift)
+}
